Create convolution.py
To introduce you to convolution, it is useful to consider a simple example. A square wave will be the signal, and a five-point linear decay function will be the kernel. Imagine taking the kernel, flipping it backward, and then dragging it forward in time so it passes over the signal. As the kernel passes over the signal, it drags on the signal, and the result of the drag is something that is neither the signal nor the kernel but looks a bit like each of them. Thus, conceptually, convolution entails sliding the kernel along with the data and obtaining a result that shows what features the signal and the kernel have in common. This “commonality” at each time point is computed as the mapping between the kernel and the signal, that is, the dot product.